# src/api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Calendly Integration
  description: Backend APIs for managing data for Calendly

basePath: /api
schemes:
  - http
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: x-api-key

paths:
  /v1/auth/authorize:
    get:
      tags:
        - "Authentication"
      summary: "Start Calendly OAuth Flow"
      description: "Redirects user to Calendly's OAuth consent screen."
      operationId: authorizeController
      produces:
        - "application/json"
      security:
        - ApiKeyAuth: []
      parameters: []
      responses:
        "302":
          description: "Redirect to Calendly OAuth"
        "500":
          description: "Server error"
      x-swagger-router-controller: "authController"

  /v1/auth/callback:
    get:
      tags:
        - "Authentication"
      summary: "Calendly OAuth Callback"
      description: "Receives the authorization code from Calendly and exchanges it for access/refresh tokens."
      operationId: callbackController
      produces:
        - "application/json"
      parameters:
        - name: code
          in: query
          description: "Authorization code returned by Calendly"
          required: true
          type: string
      responses:
        "200":
          description: "Tokens exchanged and saved successfully"
        "400":
          description: "Missing or invalid authorization code"
        "500":
          description: "Server error"
      x-swagger-router-controller: "authController"

  /v1/sync/events:
    post:
      tags:
        - "Sync"
      summary: "Sync Calendly Events"
      description: "Fetches scheduled events from Calendly, handles pagination, and stores them in the local database."
      operationId: syncEventsController
      produces:
        - "application/json"
      parameters:
        - in: body
          name: body
          required: false
          schema:
            type: object
            properties:
              pageLimit:
                type: integer
                description: "Maximum number of pages to fetch"
                default: 5
      responses:
        "200":
          description: "Events synced successfully"
        "400":
          description: "Invalid parameters"
        "500":
          description: "Server error"
      x-swagger-router-controller: "syncController"

  /v1/events:
    get:
      tags:
        - "Local Events"
      summary: "List events from local database"
      description: "Returns paginated events stored locally in MongoDB (NOT fetched from Calendly)."
      operationId: getEventsController
      security:
        - ApiKeyAuth: []
      parameters:
        - name: x-api-key
          in: header
          required: true
          type: string
        - name: page
          in: query
          type: integer
          default: 1
        - name: limit
          in: query
          type: integer
          default: 20
        - name: startDate
          in: query
          type: string
          format: date-time
        - name: endDate
          in: query
          type: string
          format: date-time
        - name: sort
          in: query
          type: string
          enum: ["asc", "desc"]
          default: "desc"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Events fetched successfully"
        "401":
          description: "Unauthorized"
        "500":
          description: "Server error"
      x-swagger-router-controller: "eventsController"

  /v1/eventsById:
    get:
      tags:
        - "Local Events"
      summary: "Get event by ID from local DB"
      description: "Fetch a single event stored locally (NOT fetched from Calendly)."
      operationId: getEventByIdController
      security:
        - ApiKeyAuth: []
      parameters:
        - name: x-api-key
          in: header
          required: true
          type: string
        - name: id
          in: query
          required: true
          type: string
          description: "MongoDB _id of the event"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Event fetched successfully"
        "401":
          description: "Unauthorized"
        "404":
          description: "Event not found"
        "500":
          description: "Server error"
      x-swagger-router-controller: "eventsController"

  /v1/webhook/register:
    post:
      tags:
        - "Webhooks"
      summary: "Register Calendly webhook subscription"
      description: "Creates a webhook subscription in Calendly for the authenticated user's organization."
      operationId: registerWebhookController
      produces:
        - "application/json"
      responses:
        "200":
          description: "Webhook subscription created successfully"
        "400":
          description: "Missing token or organization details"
        "401":
          description: "Unauthorized"
        "500":
          description: "Server error"
      x-swagger-router-controller: "webhookController"

  /v1/webhooks/receive:
    post:
      tags:
        - "Webhooks"
      summary: "Receive Calendly webhook events"
      description: "Webhook receiver endpoint for Calendly. Validates signature, logs payloads, and queues events for processing."
      operationId: receiveWebhookController
      produces:
        - "application/json"
      responses:
        "200":
          description: "Webhook received successfully"
        "400":
          description: "Invalid webhook payload"
        "401":
          description: "Invalid or missing signature"
        "500":
          description: "Server error"
      x-swagger-router-controller: "webhookController"